fix(cloudflare): keep built embeddings picklable
The executor-backed async adapter was defined inside build_embeddings, so the embeddings it returned could not be pickled (pickle and dill both raise PicklingError), unlike the Community class it replaces. The Redis cache dill-serializes vertex results and now skipped any entry holding them.

Move the adapter to lfx_bundles.cloudflare.cloudflare_common so instances pickle by reference, on both the installed-module and exec'd-component-code paths.
